Add test_base_viz
